Module google_gmailpostmastertools1_beta1::api

Structs

DeliveryError
Metric on a particular delivery error type.

Domain
A registered domain resource in the Postmaster API.

DomainGetCall
Gets a specific domain registered by the client. Returns NOT_FOUND if the domain does not exist.

DomainListCall
Lists the domains that have been registered by the client. The order of domains in the response is unspecified and non-deterministic. Newly created domains will not necessarily be added to the end of this list.

DomainMethods
A builder providing access to all methods supported on domain resources. It is not used directly, but through the PostmasterTools hub.

DomainTrafficStatGetCall
Get traffic statistics for a domain on a specific date. Returns PERMISSION_DENIED if user does not have permission to access TrafficStats for the domain.

DomainTrafficStatListCall
List traffic statistics for all available days. Returns PERMISSION_DENIED if user does not have permission to access TrafficStats for the domain.

FeedbackLoop
Feedback loop identifier information.

IpReputation
IP Reputation information for a set of IPs in a specific reputation category.

ListDomainsResponse
Response message for ListDomains.

ListTrafficStatsResponse
Response message for ListTrafficStats.

PostmasterTools
Central instance to access all PostmasterTools related resource activities

TrafficStats
Email traffic statistics pertaining to a specific date.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.